See RN6.5.1 (20260921)
Author:         cmeng
- WebViewFragment will process user backKey to go back to previous webpage if allowed.
- Ensure the webView reloads the user newly defined Web Access Page on return from Settings.
- Execute encryption of omemo message in chatRoom in background to avoid ANR.
- Make media call from recent conversion or call history call via JingleMessage protocol if supported.
- Upgrade smack to v4.5.0 release.
- - Upgrade jni libraries:
    * libvpx 1.17.0 <= 1.16.0
    * openssl 3.6.4 <= 3.5.2; and update all Hmac.c deprecated function calls.
- Upgrade android gradle build-tools to 9.3.2; change all case R.id.xxx to if-else to resolve "Constant expression required" error.

+++ aTalk/src/main/java/org/atalk/android/plugin/certconfig/TLS_Configuration.java
@@ -127,46 +127,41 @@ public class TLS_Configuration extends BaseFragment
         FragmentTransaction ft = getParentFragmentManager().beginTransaction();
         ft.addToBackStack(null);
 
-        switch (v.getId()) {
-            case R.id.cmd_add:
-                dialog = CertConfigEntryDialog.getInstance(CertificateConfigEntry.CERT_NONE, this);
+        int id = v.getId();
+        if (id == R.id.cmd_add) {
+            dialog = CertConfigEntryDialog.getInstance(CertificateConfigEntry.CERT_NONE, this);
+            dialog.show(ft, "CertConfigEntry");
+        }
+        else if (id == R.id.cmd_remove) {
+            if (mCertEntry != null) {
+                Timber.d("Certificate Entry removed: %s", mCertEntry.getId());
+                CertConfigActivator.getCertService().removeClientAuthCertificateConfig(mCertEntry.getId());
+            }
+        }
+        else if (id == R.id.cmd_edit) {
+            if (mCertEntry != null) {
+                dialog = CertConfigEntryDialog.getInstance(mCertEntry, this);
                 dialog.show(ft, "CertConfigEntry");
-                break;
-
-            case R.id.cmd_remove:
-                if (mCertEntry != null) {
-                    Timber.d("Certificate Entry removed: %s", mCertEntry.getId());
-                    CertConfigActivator.getCertService().removeClientAuthCertificateConfig(mCertEntry.getId());
-                }
-                break;
-
-            case R.id.cmd_edit:
-                if (mCertEntry != null) {
-                    dialog = CertConfigEntryDialog.getInstance(mCertEntry, this);
-                    dialog.show(ft, "CertConfigEntry");
-                }
-                break;
+            }
         }
     }
 
     @Override
     public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
         String enabled = Boolean.valueOf(isChecked).toString();
-        switch (buttonView.getId()) {
-            case R.id.cb_crl:
-                CertConfigActivator.getConfigService().setProperty(
-                        CertificateService.PNAME_REVOCATION_CHECK_ENABLED, isChecked);
+        int id = buttonView.getId();
+        if (id == R.id.cb_crl) {
+            CertConfigActivator.getConfigService().setProperty(
+                    CertificateService.PNAME_REVOCATION_CHECK_ENABLED, isChecked);
 
-                System.setProperty(CertificateService.SECURITY_CRLDP_ENABLE, enabled);
-                System.setProperty(CertificateService.SECURITY_SSL_CHECK_REVOCATION, enabled);
-                chkEnableOcsp.setEnabled(isChecked);
-                break;
-
-            case R.id.cb_ocsp:
-                CertConfigActivator.getConfigService().setProperty(
-                        CertificateService.PNAME_OCSP_ENABLED, isChecked);
-                Security.setProperty(CertificateService.SECURITY_OCSP_ENABLE, enabled);
-                break;
+            System.setProperty(CertificateService.SECURITY_CRLDP_ENABLE, enabled);
+            System.setProperty(CertificateService.SECURITY_SSL_CHECK_REVOCATION, enabled);
+            chkEnableOcsp.setEnabled(isChecked);
+        }
+        else if (id == R.id.cb_ocsp) {
+            CertConfigActivator.getConfigService().setProperty(
+                    CertificateService.PNAME_OCSP_ENABLED, isChecked);
+            Security.setProperty(CertificateService.SECURITY_OCSP_ENABLE, enabled);
         }
     }
